Moves e2e service util functions into service_util.go and cleans up test codes
