+++ test/e2e/framework/BUILD
@@ -26,6 +26,7 @@ go_library(
         "perf_util.go",
         "pods.go",
         "resource_usage_gatherer.go",
+        "service_util.go",
         "test_context.go",
         "util.go",
     ],
@@ -36,6 +37,7 @@ go_library(
         "//pkg/api:go_default_library",
         "//pkg/api/errors:go_default_library",
         "//pkg/api/v1:go_default_library",
+        "//pkg/api/v1/service:go_default_library",
         "//pkg/api/validation:go_default_library",
         "//pkg/apimachinery/registered:go_default_library",
         "//pkg/apis/apps/v1beta1:go_default_library",
@@ -82,6 +84,7 @@ go_library(
         "//pkg/util/exec:go_default_library",
         "//pkg/util/intstr:go_default_library",
         "//pkg/util/labels:go_default_library",
+        "//pkg/util/net:go_default_library",
         "//pkg/util/rand:go_default_library",
         "//pkg/util/runtime:go_default_library",
         "//pkg/util/sets:go_default_library",
